Move MemberRollController to pkg/servicemesh/controller/memberoll
