+++ pilot/pkg/config/kube/ingress/status.go
@@ -37,7 +37,7 @@ import (
 	controller2 "istio.io/istio/pilot/pkg/serviceregistry/kube/controller"
 	"istio.io/istio/pkg/listwatch"
 	queue2 "istio.io/istio/pkg/queue"
-	meshcontroller "istio.io/istio/pkg/servicemesh/controller"
+	meshcontroller "istio.io/istio/pkg/servicemesh/controller/memberroll"
 
 	"istio.io/pkg/log"
 )
