+++ pilot/pkg/serviceregistry/kube/controller/namespacecontroller.go
@@ -36,7 +36,7 @@ import (
 
 	"istio.io/istio/pkg/listwatch"
 	"istio.io/istio/pkg/queue"
-	meshcontroller "istio.io/istio/pkg/servicemesh/controller"
+	meshcontroller "istio.io/istio/pkg/servicemesh/controller/memberroll"
 	certutil "istio.io/istio/security/pkg/util"
 )
