+++ pkg/listwatch/listwatch.go
@@ -20,13 +20,13 @@ package listwatch
 import (
 	"sync"
 
+	controller "istio.io/istio/pkg/servicemesh/controller/memberroll"
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/watch"
 	"k8s.io/client-go/tools/cache"
 
-	"istio.io/istio/pkg/servicemesh/controller"
 	"istio.io/pkg/log"
 )
